* Warning fixes. * Fix TLS dir test. * Readd BOOST_USE_WINDOWS_H.

+++ src/memory/memory/pelib/section_list.cpp
@@ -113,8 +113,8 @@ SectionIterator& SectionIterator::operator++()
 
   PIMAGE_SECTION_HEADER new_base = reinterpret_cast<PIMAGE_SECTION_HEADER>(
     impl_->section_->GetBase()) + 1;
-  impl_->section_ = Section(*impl_->process_, *impl_->pe_file_, number + 1, 
-    new_base);
+  impl_->section_ = Section(*impl_->process_, *impl_->pe_file_, 
+    static_cast<WORD>(number + 1), new_base);
   
   return *this;
 }
